+++ includes/class-settings.php
@@ -94,7 +94,12 @@ class Settings {
 	 */
 	public static function save() {
 		$settings = apply_filters( 'murmurations_aggregator_save_settings', self::$settings );
-		return update_option( 'murmurations_aggregator_settings', $settings );
+		$existing = get_option( 'murmurations_aggregator_settings');
+		if( $existing !== $settings ) {
+			return update_option( 'murmurations_aggregator_settings', $settings );
+		} else {
+			return true;
+		}
 	}
 
 	/**
